Added REAL collision boxes for custom blocks!. Added rotation of the blocks. Added light factor to custom blocks. Added placeholder for custom models.

+++ src/main/java/com/volumetricpixels/rockyapi/material/Block.java
@@ -94,14 +94,14 @@ public interface Block extends Material {
 	 * 
 	 * @return
 	 */
-	public int getLightLevel();
+	public float getLightLevel();
 
 	/**
 	 * 
 	 * @param level
 	 * @return
 	 */
-	public Block setLightLevel(int level);
+	public Block setLightLevel(float level);
 
 	/**
 	 *
